feat: personnaliser l'offset de LocalInfoBox
