+++ Assets/Scripts/Interfaces/ILocalInfoBoxTarget.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 6704ab1e7b4d4f03844844ef8ec4a9ad
