+++ Assets/Scripts/MonoBehavioursUsed/NPCBase.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: a17a1a7969074f78a26ee24f56fe7f6d
